+++ R/drop_labels.R
@@ -67,6 +67,8 @@ drop_labels_helper <- function(x, drop.na) {
   tidy.labels <- get_labels(x, attr.only = T, include.values = "n", include.non.labelled = F, drop.na = T)
   # return x, if no attribute
   if (is.null(tidy.labels)) return(x)
+  # all missing in variable?
+  if (all(is.na(x))) return(x)
   # remove labels with no values in data
   tidy.labels <- tidy.labels[get_values(x) %in% names(table(x))]
   # set labels